minor fix to api-doc rendering
It's not really necessary to specify that these are optional, this should be
clear to anyone familiar with python since default values are provided -- and
the current syntax is incorrect, and results in the optional parameters not
being rendered as parameters, just as flowing text. If we really want to keep
the "The following are optional" text, an empty line needs to be added after
the text and before the parameters themselves.