syntax_wrapper: fix checkparams decorator
There can be multiple arguments that are __eq__ to the current value, we need
to use the correct index. In particular, the `__eq__` method for `VexValues`
always returns a `VexValue` which is truthy, so `args.index(arg)` will always
return `0` since the first arg is a VexValue and thus `args[0] == arg` is truthy.